Add set_options_cf
