+++ src/ops/mod.rs
@@ -70,7 +70,7 @@ pub use self::merge::{Merge, MergeCF, MergeCFOpt, MergeOpt};
 pub(crate) use self::perf::PerfInternal;
 pub use self::property::{GetProperty, GetPropertyCF};
 pub use self::put::{Put, PutCF, PutCFOpt, PutOpt};
-pub use self::set_options::SetOptions;
+pub use self::set_options::{SetOptions, SetOptionsCF};
 pub(crate) use self::snapshot::SnapshotInternal;
 pub use self::snapshot::Snapshotable;
 pub use self::write_batch::{WriteBatchWrite, WriteBatchWriteOpt};
